refactor: change directories for refactor
