+++ src/main/kotlin/blackjack/controller/BlackJackController.kt
@@ -1,6 +1,6 @@
 package blackjack.controller
 
-import blackjack.model.participant.PlayerFactory
+import blackjack.model.PlayerFactory
 import blackjack.model.Statistics
 import blackjack.view.InputView
 import blackjack.view.OutputView
