+++ src/main/kotlin/blackjack/model/Statistics.kt
@@ -1,7 +1,5 @@
 package blackjack.model
 
-import blackjack.model.participant.PlayerBackup
-
 class Statistics(val dealer: PlayerBackup, val players: List<PlayerBackup>) {
 
     val totalResult
