+++ src/main/kotlin/blackjack/model/game/CardDeck.kt
@@ -3,7 +3,7 @@ package blackjack.model.game
 import blackjack.model.card.Card
 import blackjack.model.card.Rank
 import blackjack.model.card.Suit
-import blackjack.model.participant.PlayerBackup
+import blackjack.model.PlayerBackup
 
 data class CardDeck(private val hold: Hold) {
     constructor() : this(initPokerCards())
