+++ src/test/kotlin/blackjack/model/GameManagerTest.kt
@@ -1,7 +1,6 @@
 package blackjack.model
 
 import blackjack.controller.GameManager
-import blackjack.model.participant.PlayerFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
