+++ src/test/kotlin/blackjack/model/HandTest.kt
@@ -3,7 +3,6 @@ package blackjack.model
 import blackjack.model.card.Card
 import blackjack.model.card.Rank
 import blackjack.model.card.Suit
-import blackjack.model.participant.PlayerBackup
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
