+++ src/test/kotlin/blackjack/model/HoldTest.kt
@@ -5,7 +5,6 @@ import blackjack.model.game.CardDeck
 import blackjack.model.game.Hold
 import blackjack.model.card.Rank
 import blackjack.model.card.Suit
-import blackjack.model.participant.PlayerBackup
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
