+++ src/test/kotlin/blackjack/model/PlayerFactoryTest.kt
@@ -1,6 +1,5 @@
 package blackjack.model
 
-import blackjack.model.participant.PlayerFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
